requirements.txt
setuptools>=65.5.1
wheel>=0.38.4
opencv-python==4.5.5.64
numpy==1.22.0
scikit-learn==1.0.2
Flask==2.0.3
Flask-CORS==3.0.10
pandas==1.3.5
python-dotenv==0.19.2
tensorflow==2.8.0
psycopg2-binary==2.9.3
mysql-connector-python==8.0.28
reportlab==3.6.8
Pillow==9.0.1
gunicorn==20.1.0
python-dateutil==2.8.2 